bug com um formiga parando no meio do caminho e não andando
git-svn-id: https://subversion.tecgraf.puc-rio.br/svn/svi/trunk/prototipos/antColonyOptimization@42006 67483ca6-7138-0410-9db0-fa9ffb5cce8b

+++ zhao/src/Point.cpp
@@ -10,7 +10,7 @@ Point::Point()
 
 
 
-Point::Point( float x, float y )
+Point::Point( int x, int y )
 {
     this->x = x;
     this->y = y;
